fix: Use double values

+++ src/main/java/org/ntnu/idi/idatt2106/sparesti/sparestibackend/validation/ChallengeConfigValidator.java
@@ -44,8 +44,8 @@ public class ChallengeConfigValidator extends ObjectValidator<ChallengeConfigDTO
         if (config.challengeTypeConfigs().stream()
                 .anyMatch(
                         configType ->
-                                configType.specificAmount().intValue()
-                                        > configType.generalAmount().intValue())) {
+                                configType.specificAmount().doubleValue()
+                                        > configType.generalAmount().doubleValue())) {
             throw new BadInputException("Enhetspris kan ikke være større enn ukentlig kostnad");
         }
     }
